Fix example finding
